% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADLB}
\alias{ADLB}
\title{Laboratory dataset for Glomerular Filtration Rate (GFR) measurements.}
\format{
a data frame with 13980 rows and 8 variables:
\describe{
\item{ID}{patient identifiers, numeric}
\item{TRTPN}{treatment values, 1 Active or 2 Placebo, numeric}
\item{AVAL}{measurement value, numeric}
\item{ADAY}{measurement day in the study, numeric}
\item{AVISITN}{hospital visit number, numeric}
\item{PARAM}{name of the event, GFR measurments, character}
\item{PARAMCD}{coded name of the event, GFR, character}
\item{PARAMN}{type of the event is set to 7 for all measurements, numeric}
}
}
\source{
Heerspink HL et al "Development and validation of a new hierarchical composite endpoint for clinical trials of kidney disease progression." Journal of the American Society of Nephrology (2023): \url{doi:10.1681/ASN.0000000000000243}.
}
\usage{
ADLB
}
\description{
A dataset of laboratory measurements of kidney function over time for the 1500 patients in the \code{ADSL} dataset.
}
\examples{
head(ADLB)
}
\keyword{datasets}
