% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADSL}
\alias{ADSL}
\title{Baseline characteristics dataset of patients with kidney function assessments.}
\format{
a data frame with 1500 rows and 4 variables:
\describe{
\item{ID}{patient identifiers, numeric}
\item{TRTPN}{treatment values, 1 Active or 2 Placebo, numeric}
\item{EGFRBL}{Baseline GFR values of patients, numeric}
\item{STRATAN}{strata 1-4, higher value means a higher risk for kidney disease progression, numeric}
}
}
\source{
Little, Dustin J., Gasparyan, Samvel B., et al. "Validity and utility of a hierarchical composite endpoint for clinical trials of kidney disease progression: A review." Journal of the American Society of Nephrology (2023).
}
\usage{
ADSL
}
\description{
A data frame with baseline characteristics for 1500 patients used to derive \code{KHCE} dataset.
}
\examples{
head(ADSL)
}
\keyword{datasets}
