% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hce.default.R
\name{as_hce.default}
\alias{as_hce.default}
\title{Coerce a data frame to an \code{hce} object}
\usage{
\method{as_hce}{default}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{additional parameters.}
}
\value{
an \code{hce} object.
}
\description{
Coerce a data frame to an \code{hce} object
}
\examples{
dat <- KHCE
class(dat) <- "moo" # non-existent class
as_hce(dat) # tries to convert to an hce object
## It still works because the inheritance converted it to a data frame
}
\seealso{
\code{\link[=as_hce]{as_hce()}}, \code{\link[=as_hce.data.frame]{as_hce.data.frame()}}.
}
