% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_hce.R
\name{as_hce}
\alias{as_hce}
\title{A generic function for coercing data structures to \code{hce} objects}
\usage{
as_hce(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{additional parameters.}
}
\value{
an \code{hce} object.
}
\description{
A generic function for coercing data structures to \code{hce} objects
}
\examples{
### data frames 
data(HCE1)
HCE <- as_hce(HCE1)
calcWINS(HCE)
}
\seealso{
\code{\link[=as_hce.data.frame]{as_hce.data.frame()}}, \code{\link[=as_hce.default]{as_hce.default()}}.
}
