% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWINS.R
\name{calcWINS}
\alias{calcWINS}
\title{A generic function for calculating win statistics}
\usage{
calcWINS(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame containing calculated values.
}
\description{
A generic function for calculating win statistics
}
\seealso{
\code{\link[=calcWINS.hce]{calcWINS.hce()}}, \code{\link[=calcWINS.formula]{calcWINS.formula()}}, \code{\link[=calcWINS.data.frame]{calcWINS.data.frame()}}  methods.
}
