% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWO.R
\name{calcWO}
\alias{calcWO}
\title{A generic function for calculating win odds}
\usage{
calcWO(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame containing calculated values.
}
\description{
A generic function for calculating win odds
}
\seealso{
\code{\link[=calcWO.formula]{calcWO.formula()}}, \code{\link[=calcWO.hce]{calcWO.hce()}}, \code{\link[=calcWO.data.frame]{calcWO.data.frame()}}  methods.
}
