% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaWO.R
\name{deltaWO}
\alias{deltaWO}
\title{A generic function for calculating win odds based on a threshold}
\usage{
deltaWO(x, delta, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{delta}{a numeric threshold.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame containing calculated values.
}
\description{
A generic function for calculating win odds based on a threshold
}
\seealso{
\code{\link[=deltaWO.adhce]{deltaWO.adhce()}} method.
}
