% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulae.R
\name{formulae}
\alias{formulae}
\alias{as_formulae}
\alias{as_formulae.formula}
\title{Decompose formula objects}
\usage{
as_formulae(x, ...)

\method{as_formulae}{formula}(x, ...)
}
\arguments{
\item{x}{a formula object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list with standardised components named \code{AVAL}, \code{TRTP},
\code{COVAR}, and \code{GROUP} where present, containing the original variable
names extracted from \code{x}, plus a reconstructed formula.
}
\description{
Decompose formula objects to treatment, response, covariate, and grouping variables using standardized names
\code{TRTP}, \code{AVAL}, \code{COVAR}, and \code{GROUP} respectively.
}
\examples{
# Example 1: All variables present
as_formulae(x = visit4 ~ treatment + age | sex)
# Example 2: Extra covariates, no grouping variable will drop the extra variable
as_formulae(x = visit4 ~ treatment + age + sex)
# Example 3: When more than two grouping variables, the last one will be used.
 as_formulae(x = visit4 ~ treatment + age | sex |site)
}
\seealso{
\code{\link[=summaryWO.formula]{summaryWO.formula()}}, \code{\link[=stratWO.formula]{stratWO.formula()}} for use.
}
