% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hce.R
\name{new_hce}
\alias{new_hce}
\title{Constructor function for the hce object}
\usage{
new_hce(x = data.frame())
}
\arguments{
\item{x}{a data frame.}
}
\value{
an object of class hce.
}
\description{
Constructor function for the hce object
}
\examples{
data(HCE1)
dat <- new_hce(x = HCE1)
class(dat)
calcWO(dat)
}
\seealso{
\code{\link[=hce]{hce()}}, \code{\link[=validate_hce]{validate_hce()}}  for the helper and validator functions of hce.
}
