% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot.hce_results}
\alias{plot.hce_results}
\title{A print method for \code{hce_results } objects}
\usage{
\method{plot}{hce_results}(x, ...)
}
\arguments{
\item{x}{an object of class \code{hce_results}.}

\item{...}{additional arguments to be passed to \code{\link[base:plot]{base::plot()}} function.}
}
\value{
no return value, called for plotting.
}
\description{
A print method for \code{hce_results } objects
}
\examples{
WO <- minWO(N = 100:1000)
plot(WO)
POW <- powerWO(N = 100:1000, WO = 1.2)
plot(POW, ylim = c(0, 1))
}
