% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_results.R
\name{print.hce_results}
\alias{print.hce_results}
\title{A print method for \code{hce_results} objects}
\usage{
\method{print}{hce_results}(x, ...)
}
\arguments{
\item{x}{an object of class \code{hce_results}.}

\item{...}{additional arguments to be passed to \code{\link[base:print]{base::print()}} function.}
}
\value{
no return value, called for printing.
}
\description{
A print method for \code{hce_results} objects
}
\examples{
print(powerWO(N = 1000, WO = 1.2))
}
