% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regWO.R
\name{regWO}
\alias{regWO}
\title{A generic function for win odds regression}
\usage{
regWO(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame containing calculated values.
}
\description{
A generic function for win odds regression
}
\seealso{
\code{\link[=regWO.data.frame]{regWO.data.frame()}}  methods.
}
