% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simNORM.R
\name{simNORM}
\alias{simNORM}
\title{Simulate multivariate normal distribution using compound symmetry covariance structure}
\usage{
simNORM(n, mean, sd = 1, rho = 0)
}
\arguments{
\item{n}{number of observations.}

\item{mean}{vector of means.}

\item{sd}{vector of length 1 of standard deviation.}

\item{rho}{vector of length 1 of between subject correlation.}
}
\value{
a matrix with \code{n} rows and \code{length(mean)} columns.
}
\description{
Simulate multivariate normal distribution using compound symmetry covariance structure
}
\examples{
MN <- simNORM(n = 1000, mean = c(1, 2), sd = 2, rho = -0.5)
colMeans(MN)
var(MN)
}
