% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratWO.formula.R
\name{stratWO.formula}
\alias{stratWO.formula}
\title{Stratified win odds with adjustment using formula}
\usage{
\method{stratWO}{formula}(x, data, ...)
}
\arguments{
\item{x}{an object of class formula.}

\item{data}{a data frame.}

\item{...}{additional parameters.}
}
\value{
a data frame containing the following columns:
\itemize{
\item WO stratified (or adjusted/stratified) win odds.
\item LCL lower confidence limit for adjusted (or adjusted/stratified) WO.
\item UCL upper confidence limit for adjusted (or adjusted/stratified) WO.
\item SE standard error of the adjusted (or adjusted/stratified) win odds.
\item WOnull win odds of the null hypothesis (specified in the \code{WOnull} argument).
\item alpha two-sided significance level for calculating the confidence interval (specified in the \code{alpha} argument).
\item Pvalue p-value associated with testing the null hypothesis.
\item WP adjusted (or adjusted/stratified) win probability.
\item LCL_WP lower confidence limit for adjusted (or adjusted/stratified) WP.
\item UCL_WP upper confidence limit for adjusted (or adjusted/stratified) WP.
\item SE_WP standard error for the adjusted (or adjusted/stratified) win probability.
\item SD_WP standard deviation of the adjusted (or adjusted/stratified) win probability.
\item N total number of patients in the analysis.
\item Type "STRATIFIED" or "STRATIFIED/ADJUSTED" depending on whether \code{COVAR} is specified.
}
}
\description{
Stratified win odds with adjustment using formula
}
\examples{
# Example 1: Stratified win odds with covariate adjustment
stratWO(x = AVAL ~ TRTP + EGFRBL | STRATAN, data = KHCE)
# Example 2: Stratified win odds without covariate adjustment
stratWO(x = AVAL ~ TRTP | STRATAN, data = KHCE)
# Example 3: Only covariate adjustment (without stratification) is implemented in regWO()
regWO(x = AVAL ~ TRTP + EGFRBL, data = KHCE)
# Example 4: Non-standardized names
dat <- data.frame(response = KHCE$AVAL, treatment = KHCE$TRTP, 
covariate = KHCE$EGFRBL, group = KHCE$STRATAN)
stratWO(x = response ~ treatment + covariate | group, data = dat)
}
\references{
Gasparyan SB et al. (2021) "Adjusted win ratio with stratification: calculation methods and interpretation." Statistical Methods in Medical Research 30.2: 580-611. \url{doi:10.1177/0962280220942558}.
}
\seealso{
\code{\link[=stratWO]{stratWO()}}, \code{\link[=stratWO.data.frame]{stratWO.data.frame()}}.
}
