% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryWO.R
\name{summaryWO}
\alias{summaryWO}
\title{A generic function for summarizing win odds}
\usage{
summaryWO(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a data frame containing calculated values.
}
\description{
A generic function for summarizing win odds
}
\seealso{
\code{\link[=summaryWO.hce]{summaryWO.hce()}}, \code{\link[=summaryWO.formula]{summaryWO.formula()}}, \code{\link[=summaryWO.data.frame]{summaryWO.data.frame()}}  methods.
}
