% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valicate_hce.R
\name{validate_hce}
\alias{validate_hce}
\title{Validator function for the hce object}
\usage{
validate_hce(x)
}
\arguments{
\item{x}{an object used for the hce validation.}
}
\value{
a validated data frame that can be used for creating an hce object.
}
\description{
Validator function for the hce object
}
\examples{
data(HCE1)
validate_hce(HCE1)
}
\seealso{
\code{\link[=hce]{hce()}}, \code{\link[=new_hce]{new_hce()}}  for the helper and constructor functions of hce.
}
