% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_hce.R
\name{new_hce}
\alias{new_hce}
\title{Constructor function for \code{hce} objects}
\usage{
new_hce(x = data.frame())
}
\arguments{
\item{x}{a data frame.}
}
\value{
an object of class \code{hce}.
}
\description{
Constructor function for \code{hce} objects
}
\examples{
data(HCE1)
dat <- new_hce(x = HCE1)
class(dat)
calcWO(dat)
}
\seealso{
\code{\link[=hce]{hce()}}, \code{\link[=validate_hce]{validate_hce()}}  for the helper and validator functions of \code{hce} objects.
}
