% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rweibullGF.R
\name{rweibullGF}
\alias{rweibullGF}
\title{Simulate random numbers from a Weibull distribution with gamma frailty}
\usage{
rweibullGF(n, shape, rate, theta = 1)
}
\arguments{
\item{n}{number of observations.}

\item{shape}{shape parameter of the Weibull distribution.}

\item{rate}{rate parameter of the Weibull distribution.}

\item{theta}{Variance parameter of the gamma frailty distribution, with a default value of 1. A value of 0 corresponds to no frailty, that is, a standard Weibull distribution.}
}
\value{
a vector of random numbers of length \code{n}.
}
\description{
Simulate random numbers from a Weibull distribution with gamma frailty
}
\details{
The survival function is given by \deqn{S(t|\lambda, \alpha, \gamma) = e^{-\gamma\cdot\lambda\cdot t^\alpha},\ \ \alpha>0,\,\lambda>0,} where \eqn{\alpha} is the shape parameter, \eqn{\lambda} is the rate parameter, and \eqn{\gamma} is the frailty generated from a gamma distribution with \eqn{shape = \frac{1}{\theta}}
and \eqn{scale=\theta} which results in a frailty variance of \eqn{\theta}. When \eqn{\theta = 0}, the distribution reduces to a standard Weibull distribution without frailty.

The marginal (unconditional) survival function of the Weibull distribution with gamma frailty can be expressed as \deqn{S(t|\lambda, \alpha, \theta) = \left(1 + \theta\cdot\lambda\cdot t^\alpha\right)^{-\frac{1}{\theta}},\ \ \alpha>0,\,\lambda>0,\,\theta\geq 0.}
}
\examples{
# Simulate 10 random numbers
set.seed(123)
rweibullGF(n = 10, shape = 2, rate = 0.5, theta = 0.1)
}
\references{
Wienke A. ``Frailty Models in Survival Analysis." Chapman and Hall/CRC (2010).
}
\seealso{
\code{\link[=simTTE]{simTTE()}}  for simulation of a two-event time-to-event endpoint from this distribution under the illness-death model.
}
