% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratWO.R
\name{stratWO}
\alias{stratWO}
\title{A generic function for stratified win odds with adjustment}
\usage{
stratWO(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list containing the stratified results and results by strata.
}
\description{
A generic function for stratified win odds with adjustment
}
\seealso{
\code{\link[=stratWO.data.frame]{stratWO.data.frame()}}  methods.
}
